Cleanup in helper traits for Collections
Some helper subtraits are defined to abstract some duplicated boilerplate, but the current
code did not do so in a consistent way.
Now we have:
* ChildlessExp[T] extends Exp[T] //for nodes without children

* QueryOp[T] extends QueryReifier[T] //interpretation gives the same node; it's
for

* ChildlessQueryReifier[T] extends ChildlessExp[QueryReifier[T]] with QueryOp[T]
//Both things together.

= Other changes =

We had QueryOp, ConstBase and ChildlessQueryReifier, defined at different places
and with unnecessary duplication between them, which was removed. Now ConstBase
becomes instead ChildlessExp.

* Queryable used ConstBase, and the implementation of interpret was quite
  roundabout.
* CollectionReifier duplicated this boilerplate
* View and Force inherited directly and indirectly from QueryReifier, fixed.