QueryableTest: experiment with collection definition and start test for IncrementalSet
First, extend existing code: create IncArrayBuffer mixing in all the appropriate traits.
Second, try writing a test for IncrementalSet. However, it is not usable at all, since it does not offer a set interface - tons of problem are remarked in testIncremental() as commented out code. To fix this, one should forward all methods to the underlying collection; alternatively, one should make IncrementalSet a mixin. The second solution is better, since we don't want to have wrapper objects with different identities.