OpenEncoding: try to extend it to full functionality
I modularized the internal content (by splitting it into different traits) and started implementing TraversableExp, analogous to NumExp. Its implementation is definitely more complex and reveals some unexpected problems in comparison - namely, how to ensure that NumExp[T] instead of Exp[T] is deduced as the type of the passed closure. At least, we don't have the problem that T cannot be deduced.
This commit also starts solving the given problem by saving into each TraversableExp an instance of CanBuildExp, and having ExpT as a type parameter of TraversableExp.
Next, we must use ExpT in signatures and bodies of members of TraversableExp.