Seemingly successful encoding of operations on Maps
Various failed attempts are also included and will be removed. I adapted classes handling
pairs to write appropriate examples.

Main goal here: encode different versions of MapOp, FlatMap and WithFilter,
accepting (implicit c: CanBuildFrom[Repr, B, That]), like corresponding
operations in Scala collections, and extending That.
1) TraversableOpsExpressionTreeV2 attempts to abstract over
the source collection, accepting it as a higher-kinded type parameter
of kind * -> *. However Map has kind * -> * -> *, so this attempt fails.
2) The next attempt is to define MapOp etc. within the source object,
where the original instantiation is known, so that higher-kinded
abstractions are not needed. This is identical to the definition of
nodes which reify operations on views in Scala's TraversableView,
IterableView, and so on. This attempt works quite well; only
the definition of WithFilter must be duplicated.