Reduce type annotations in Optimization
It is notable (by observing the typechecker output) that the type
annotation to buildJoinTyped is not automatically deduced.
Type inference is always free (according to specs) to choose any
suitable solution, so this is not a bug according to the spec.
Moreover, given that the types which should be deduced are
only type tokens anyway, it's not even clear what'd be the
point.