Reify T in Exp[T] through ClassManifest
That's useful to match on the specific type of an expression
at runtime in the optimizer. That's required e.g. for creating
join nodes - that is not possible if the source object is of type
FilterMonadic.
This requires massive surgery to pass around the needed implicit
ClassManifests.

Meanwhile:
- remove duplication between SimpleOpenEncoding and Lifting
- FuncExp becomes invariant for now - I'll try to fix that later.
- More classes start using ExpNodeHelpers; the reason is that it
automatically provides manifests, but it also makes
genericConstructor safer (checks the arity of the passed argument
vector) and centralizes type-unsafe code.