Complete opaltests.BasicTests by runtime type-testing
Now extractors check the runtime type of the passed expression by
calling interpret on it. The computed result is currently thrown
away and recomputed, to keep using liftCall and creating nodes.
Otherwise, we'd have to compute all calls directly.

The current code is not yet in to-be-generated form - before we need
agree on what to do.