Fix IVM test errors (excluding ops on Arrays)
This is still done with the old notification setup (using
FuncExp.interpretHook and so on), but at least fixes the error.
To fix the problems, we ensure in FlatMapMaintainerExp to call
f(v).startListening on both initial values (in the constructor) and values
we get from notifications (which was never done before, I don't know why and
how it worked).