BasicTests: Use a class instead of a type synonym
Given that type inference cannot infer type synonyms (I assume), this is the
only solution I currently see. For it to be practical, much of the code
should be generated - which doesn't sound easy.

The previous code just required declaring and using SND in one place.
That might be improved using Scalaz' partial type application functions:
http://scalaz.googlecode.com/svn/continuous/latest/browse.sxr/scalaz/PartialApplys.scala.html