Use an environment instead of substitution
Typing becomes more messy, now that TypedVar nodes need a type. I will clean
it up a bit later, but it is not possible to do much better.
Furthermore, using an environment seems to be as fast as HOAS, and faster
after optimization. Therefore, we might as well avoid using FuncExp and
convert immediately from HOAS to FOAS.

TODO: make the environment thread-local.