Introduce support for let-bindings in our for comprehensions
Normal variable bindings of the form:
for (a <- collection; letVar = valueToLet) yield bar
for collection: Exp[Traversable[A]] are desugared to
for ((a, letVar) <- v) yield bar
for some v: Exp[(A, B)]; we cannot support this pattern matching without
altering the Scala compiler.
An alternative which work in normal comprehensions is to use:

for (a <- collection; letVar <- Seq(valueToLet)) yield bar

But alas, this does not work for smart collections, since it forms
a Seq[Exp[B]].

Finally, I came up with a working variant, which is included.
Next commit will make it even simpler!