Simplify implementation of Let
In the previous commit, Let.withFilter relied on Option and OptionOps.
I tried therefore to use Some directly, but it didn't work, because
using Some instead of Let created a value of type Option[Exp[T]].

We need instead Exp[Option[T]], which we produce simply with
onExp(v)('Some, Some(_)).

Btw, this code is also more robust - the presence of Let hid the lack of
Option.withFilter before.