Optimize coll.size > 0 into coll.nonEmpty (and similars)
This optimization is the first matching Call* nodes.
The code compiles, but there is no typechecking going on - the symbol
name does not allow deducing the types of the call arguments. Moreover,
should symbols be renamed, there is no automated support for renaming
references.
This optimization is yet neither tested, nor activated.