Optimization: fix ambiguity between transformers and methods
Some transformers (like mergeMaps) are recursive, but they should
reinvoke the transfomer itself, and not the facade method from
Optimization which invokes exp.transform - that would perform an extra
pass on the whole subtree, including already processed parts.

To avoid ambiguity, split Optimization into two different objects.
The ambiguity was introduced when merging the Optimization class and
object - the separation was needed, having a class for one part not.