Start rewriting anti-joins using lets, expose bug caused by dynamic scoping
This commit rewrites anti-joins without building a custom node -
we instead produce a complex expression tree which has the right
behavior under interpretation.

To do this, we need to build an index on the fly and reuse it -
we need a call-by-value let to do this, and we build let expressions
using the translation to lambda-expressions (which is valid since our
expression language does not support let-polymorphism and this
desugaring happens after type assignment).

The test case, however, produces lazy views which store interpreted
FuncExpInt - which do not work because I implemented dynamic scoping by
accident, instead of static scoping. This will be fixed in a future
commit.