Fix infinite recursion in CNF
The old code was just wrong in the last case - it couldn't handle the
new test, which computes cnf((a and b) or (c and d)).
The failure was triggered by optimizations in FindBugsAnalyses
(for the "UUF_UNUSED_FIELD Los" query).
The new code was cross-checked with TypeChef's implementation
(both results and their source) and with (a selection of) the existing
tests.