Make expression-tree nodes even more typesafe
Now unary and binary ExpNodeHelpers require a class to declare its
own type (which is checked) - this allows checking statically that
copy returns the correct type, rather than requiring a test to discover
that or seeing expressions change upon optimization.

This is something to remark in the paper.

This had been tested in BinaryOpTrait2, which is now merged into
BinaryOpTrait.