Fix bug due to mixture of FuncExp and FuncExpInt
Apparently, we need to use only FuncExpInt, and this commit does exactly
that. Therefore, no FuncExp objects are built any more.

The bug showed up after
Optimization.hoistFilter produced a FuncExp (which does not save its
environment in interpret()) containing a variable which was closed by
an outer FuncExpInt: the inner FuncExp in practice behaved as if defined
with dynamic scoping, which caused a crash in OptimTests because enough
laziness was involved thanks to views (as usual).